Add tutorial files (#100)
* Add tutorial files

Tutorial files that were previously kept in another repository are added
locally. The fsh-tutorial directory contains these files.

Add step to the github action for deploying the site that rebuilds the
tutorial zip.

* Add script for rebuilding tutorial zip

Add section to README about rebuilding the tutorial zip. Use the script
in the deployment action.

Tutorial zip is rebuilt, but there are no changes to the contents. I
think there may be some nondeterministic components in this file.